+++ CHANGELOG.md
@@ -63,3 +63,8 @@ node server/migrate/migratetov0.5.2.js
 ## v0.5.6rc1
 
 - no longer redirects the users after avatar upload
+
+## v0.5.6rc2
+
+- support account verification in one week
+- support resend verification email